requirements.txt
boto3==1.38.32
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
networkx==3.5
PyMuPDF==1.26.0
python-docx==1.1.2
python-dotenv==1.1.0